Move Migration Part to a Single File
