+++ runtime/crab/src/migrations.rs
@@ -1,0 +1,28 @@
+// --- paritytech ---
+#[allow(unused)]
+use frame_support::{migration, traits::OnRuntimeUpgrade, weights::Weight};
+// --- darwinia-network ---
+#[allow(unused)]
+use crate::*;
+
+fn migrate() -> Weight {
+	0
+	// RuntimeBlockWeights::get().max_block
+}
+
+pub struct CustomOnRuntimeUpgrade;
+impl OnRuntimeUpgrade for CustomOnRuntimeUpgrade {
+	#[cfg(feature = "try-runtime")]
+	fn pre_upgrade() -> Result<(), &'static str> {
+		Ok(())
+	}
+
+	#[cfg(feature = "try-runtime")]
+	fn post_upgrade() -> Result<(), &'static str> {
+		Ok(())
+	}
+
+	fn on_runtime_upgrade() -> Weight {
+		migrate()
+	}
+}
